add routing to ListItems
